llm_chatbot/pyproject.toml
[project]
name = "llm-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "gradio>=5.29.0",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.0",
]
